Fixes:
Modified pkg/plugin/httproute.go and pkg/plugin/grpcroute.go:

Added logic to the setHTTPRouteWeight and setGRPCRouteWeight functions to enable it to read managedRouteMap in ConfigMap.
managedRouteMap identifies which routing rules were created by setHeaderRoute (i.e., Header routes).
Skip these Header routing rules when updating weights. This means that the setWeight operation will only affect the main routing rule and will not override the weight of the Header routing rule.
The result is that the weights of the Header routing rules will remain at their default values (i.e., nil, which in the Gateway API means that 100% of traffic is directed to that backend), achieving what you would expect from "100% of requests with Header are routed to the new version".
Updated related tests:

Modified test/e2e/singleheaderbasedhttproutetest.go and test/e2e/singleheaderbasedgrpcroutetest.go.
The test logic no longer expects the weight of the Header route to change with the value of setWeight, but verifies that its weight remains nil (the default).
Now, when you configure setWeight: 0 and setHeaderRoute, the weight of the Header route in the resulting routing rule will no longer be forced to be 0, but will remain valid, correctly directing the matching traffic to the Canary service.
思考中:

------------
Fixes:

Modified pkg/plugin/httproute.go and pkg/plugin/grpcroute.go:

Added logic in the setHTTPRouteWeight and setGRPCRouteWeight functions to read the managedRouteMap from the ConfigMap.
Used managedRouteMap to identify which route rules were created